requirements.txt
dependency_injector==4.41.0
python-dotenv==1.0.0
redis==3.5.3
pyclean==2.7.0